% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{commonCNV}
\alias{commonCNV}
\title{A SummarizedExperiment containing
TCGA data: DNA methylation platform 450K chromossome 9
for 10 LGG samples and 10 GBM samples}
\format{
A data frame with 7454 rows and 6 columns
}
\description{
A SummarizedExperiment containing
TCGA data: DNA methylation platform 450K chromossome 9
for 10 LGG samples and 10 GBM samples
}
\examples{
data("CNV.hg19.bypos.111213")
}
\keyword{internal}
