% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{exp_gbm}
\alias{exp_gbm}
\title{A gene expression matrix with 20 GBM samples}
\format{
Gene expression: A SummarizedExperiment object
 with 21022 rows and 20 columns
}
\description{
A gene expression matrix with 20 GBM samples
}
\examples{
data("TCGA_GBM_Transcriptome_20_samples")
}
\keyword{internal}
