% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{exp_lgg}
\alias{exp_lgg}
\title{A gene expression matrix with 20 LGG samples}
\format{
Gene expression: A SummarizedExperiment object
with 21022 rows and 20 columns
}
\description{
A gene expression matrix with 20 LGG samples
}
\examples{
data("TCGA_LGG_Transcriptome_20_samples")
}
\keyword{internal}
