% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{GBMmut}
\alias{GBMmut}
\title{GBM GDC maf files for  Somatic Mutation Calling Workflow mutect2}
\format{A matrix with 72102 rows and 114 columns}
\description{
GBM GDC maf files for  Somatic Mutation Calling Workflow mutect2
}
\examples{
data("mafMutect2LGGGBM")
}
\keyword{internal}
