% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{genes}
\alias{genes}
\title{A data frame object with gene information (hg19)}
\format{A dataframe object}
\description{
A data frame object with gene information (hg19)
}
\examples{
data("genes_GR")
}
\keyword{internal}
