% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{genes_GR}
\alias{genes_GR}
\title{A GRanges object with gene information (hg19)}
\format{A GRanges object}
\description{
A GRanges object with gene information (hg19)
}
\examples{
data("genes_GR")
}
\keyword{internal}
