% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{gistic.allbygene}
\alias{gistic.allbygene}
\title{A subset of GBM GISTIC2 results, which is used to identify genes targeted by somatic copy-number alterations (SCNAs)
From: GDAC firehose, downloaded with RTCGAtoolbox}
\format{A matrix with 24776 rows and 580 columns}
\description{
A subset of GBM GISTIC2 results, which is used to identify genes targeted by somatic copy-number alterations (SCNAs)
From: GDAC firehose, downloaded with RTCGAtoolbox
}
\examples{
data("GBMGistic")
}
\keyword{internal}
