% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{histone.marks}
\alias{histone.marks}
\title{histone marks specific for brain tissue from the Roadmap database.}
\format{A matrix with 72102 rows and 114 columns}
\description{
histone marks specific for brain tissue from the Roadmap database.
}
\examples{
data("histoneMarks")
}
\keyword{internal}
