% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{LGGmut}
\alias{LGGmut}
\title{LGG  GDC maf files for  Somatic Mutation Calling Workflow mutect2}
\format{A matrix with 29840 rows and 114 columns}
\description{
LGG  GDC maf files for  Somatic Mutation Calling Workflow mutect2
}
\examples{
data("maf_mutect2_LGG_GBM")
}
\keyword{internal}
