% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{maf}
\alias{maf}
\title{Merged LGG and GBM GDC MAF files from GDC
workflow: Aliquot Ensemble Somatic Variant Merging and Masking}
\format{
A matrix with 87957 rows and 141 columns
}
\description{
Merged LGG and GBM GDC MAF files from GDC
workflow: Aliquot Ensemble Somatic Variant Merging and Masking
}
\examples{
data("maf_lgg_gbm")
}
\keyword{internal}
