% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{mut}
\alias{mut}
\title{Merged LGG and GBM GDC maf files for  Somatic Mutation Calling Workflow mutect2}
\format{A matrix with 101942 rows and 114 columns}
\description{
Merged LGG and GBM GDC maf files for  Somatic Mutation Calling Workflow mutect2
}
\examples{
data("mafMutect2LGGGBM")
}
\keyword{internal}
