% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{package}
\name{TCGAWorkflowData}
\alias{TCGAWorkflowData}
\alias{TCGAWorkflowData-package}
\title{Data for TCGA Workflow}
\description{
This experimental data package has the data necessary to follow the
TCGA Workflow: Analyze cancer genomics and epigenomics data using Bioconductor packages.
It contains the following files:
\itemize{
  \item GBMnocnvhg19: segmented CNV from SNP array (Affymetrix Genome-Wide Human SNP Array
    6.0) for 20  Glioblastoma multiforme (GBM)  samples.
  \item GBMIllumina_HiSeq: a matrix with  raw expression signal for expression of a gene for
    20 GBM (Glioblastoma multiforme) samples
  \item LGGIllumina_HiSeq: a matrix with  raw expression signal for expression of a gene for
    20 LGG (Lower grade glioma) samples
  \item met20SamplesGBMLGGchr9:  DNA methylation matrix from Infinium HumanMethylation450 platform
    for 10 LGG (Lower grade glioma) and 10 GBM (Glioblastoma multiforme).
    It has only probes in chromossome 9  in order to make the example  of the workflow faster
  \item elmerExample: Contains a DNA methylation matrix (only probes in chromossome 9) and a gene expression matrix for
    10  LGG (Lower grade glioma) and 10 GBM (Glioblastoma multiforme) in the required format
    for to execute the R/Bioconductor ELMER package analysis and a vector
    identifying which sample belongs to each tumor type.
   \item GBMGistic: a subset of GBM GISTIC results which is used to identify genes targeted by
     somatic copy-number  alterations (SCNAs)
   \item biogrid: biogrid information
  \item mafMutect2LGGGBM: Mutation annotation files for LGG (Lower grade glioma) and GBM (Glioblastoma multiforme)
    samples merged into a single matrix. The GDC Somatic Mutation Calling Workflow mutect2 was used to create this MAF files.
  \item markersMatrix: Probes meta file from broadinstitute website for Copy Number Variation Analysis (CNV) analysis
  \item  histoneMarks: histone marks specific for brain tissue using from Roadmap database.
  \item  genes_GR: A GRanges Object and a dataframe  with gene information (hg19) downloaded from ENSEMBLE database using
  biomart via TCGAbiolinks
}
    For more information how to create these objects please read the vignette of this package with the
    follwing command: \code{browseVignettes("TCGAWorkflowData")}
}
\examples{
data("elmerExample")
data("GBMnocnvhg19")
data("GBMIllumina_HiSeq")
data("LGGIllumina_HiSeq")
data("met20SamplesGBMLGGchr9")
data("mafMutect2LGGGBM")
data("markersMatrix")
data("biogrid")
data("GBMGistic")
data("histoneMarks")
data("genes_GR")
}
\keyword{utilities}
