% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAWorkflowExampleData.R
\docType{data}
\name{tmp.biogrid}
\alias{tmp.biogrid}
\title{Biogrid information}
\format{Two matrices with 24776 rows and 580 columns}
\description{
Biogrid information
}
\examples{
data("biogrid")
}
\keyword{internal}
